requirements.txt
Flask
Flask-SQLAlchemy
pillow
requests
waitress
werkzeug
